+++ src/components/UserAvatar/UserAvatar.js
@@ -35,7 +35,7 @@ const StyledSpinnerGravatar = styled.div`
     background-color: ${props => props.theme.lightDarker};
 `;
 
-const UserAvatar = ({ userId, size }) => {
+const UserAvatar = ({ userId, size, appendToTooltip }) => {
     const [contributor, setContributor] = useState(null);
     const [isLoadingContributor, setIsLoadingContributor] = useState(true);
 
@@ -54,12 +54,12 @@ const UserAvatar = ({ userId, size }) => {
     }, [userId]);
 
     return (
-        <div>
+        <>
             {userId && (
                 <Tippy
                     offset={[0, 10]}
                     placement="bottom"
-                    content={contributor?.display_name}
+                    content={`${contributor?.display_name}${appendToTooltip}`}
                     disabled={!userId || !contributor || isLoadingContributor}
                 >
                     <Link to={reverse(ROUTES.USER_PROFILE, { userId: userId })}>
@@ -74,17 +74,19 @@ const UserAvatar = ({ userId, size }) => {
                     </Link>
                 </Tippy>
             )}
-        </div>
+        </>
     );
 };
 
 UserAvatar.propTypes = {
     userId: PropTypes.string,
-    size: PropTypes.number
+    size: PropTypes.number,
+    appendToTooltip: PropTypes.string
 };
 
 UserAvatar.defaultProps = {
-    size: 28
+    size: 28,
+    appendToTooltip: ''
 };
 
 export default UserAvatar;