+++ src/constants/graphSettings.js
@@ -67,6 +67,7 @@ export const CLASSES = {
     COMPARISON: 'Comparison',
     FEATURED_COMPARISON: 'FeaturedComparison',
     FEATURED_COMPARISON_CATEGORY: 'FeaturedComparisonCategory',
+    FEATURED_COMPARISON_HOME_PAGE: 'FeaturedComparisonHomePage',
     TEMPLATE: 'ContributionTemplate',
     PREDICATE: 'Predicates',
     CLASS: 'Classes',
@@ -95,8 +96,10 @@ export const MISC = {
     UNKNOWN_ID: '00000000-0000-0000-0000-000000000000'
 };
 
+/* entities usually are returned by the backend (in _class) to specify the type of node  */
 export const ENTITIES = {
     CLASS: 'class',
     PREDICATE: 'predicate',
-    RESOURCE: 'resource'
+    RESOURCE: 'resource',
+    LITERAL: 'literal'
 };