+++ src/pages/Resources/Resource.js
@@ -1,5 +1,5 @@
 import { useState, useEffect, useRef, useCallback } from 'react';
-import { Container, Button, FormGroup, Label, FormText, ButtonGroup } from 'reactstrap';
+import { Container, Button, FormGroup, Label, FormText, ButtonGroup, Alert } from 'reactstrap';
 import { getClassById } from 'services/backend/classes';
 import { updateResourceClasses as updateResourceClassesNetwork } from 'services/backend/resources';
 import { getResource } from 'services/backend/resources';
@@ -105,6 +105,7 @@ function Resource(props) {
     const isCurationAllowed = useSelector(state => state.auth.user?.isCurationAllowed);
     const showDeleteButton = editMode && isCurationAllowed;
     const [hasObjectStatement, setHasObjectStatement] = useState(false);
+    const [hasDOI, setHasDOI] = useState(false);
     const { deleteResource } = useDeleteResource({ resourceId, redirect: true });
     const [canEdit, setCanEdit] = useState(false);
     const [createdBy, setCreatedBy] = useState(null);
@@ -141,7 +142,8 @@ function Resource(props) {
                                 getStatementsBySubjectAndPredicate({ subjectId: props.match.params.id, predicateId: PREDICATES.HAS_DOI }).then(st => {
                                     if (st.length > 0) {
                                         setIsLoading(false);
-                                        setCanEdit(false);
+                                        setHasDOI(true);
+                                        setCanEdit(isCurationAllowed);
                                     } else {
                                         setIsLoading(false);
                                         if (env('PWC_USER_ID') === responseJson.created_by) {
@@ -169,7 +171,7 @@ function Resource(props) {
                 });
         };
         findResource();
-    }, [location, props.match.params.id, resourceId]);
+    }, [location, props.match.params.id, resourceId, isCurationAllowed]);
 
     useEffect(() => {
         setCanBeDeleted((values.allIds.length === 0 || properties.allIds.length === 0) && !hasObjectStatement);
@@ -279,7 +281,11 @@ function Resource(props) {
                             )}
                         </ButtonGroup>
                     </Container>
-
+                    {editMode && hasDOI && (
+                        <Alert className="container" color="danger">
+                            This resource should not be edited because it has a published DOI, please make sure that you know what are you doing!
+                        </Alert>
+                    )}
                     {editMode && canEdit && (
                         <EditModeHeader className="box rounded-top">
                             <Title>